backend/requirements.txt
# Django and REST Framework
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Async support
uvicorn[standard]>=0.27,<1.0
gunicorn>=21.0,<23.0

# MongoDB
motor>=3.3,<4.0
pymongo>=4.6,<5.0

# Data validation
pydantic>=2.5,<3.0
pydantic-settings>=2.1,<3.0

# HTTP client
httpx>=0.26,<1.0

# AI
anthropic>=0.39,<1.0

# Environment variables
python-dotenv>=1.0,<2.0

# Development
pytest>=7.4,<9.0
pytest-asyncio>=0.23,<1.0
pytest-django>=4.7,<5.0
black>=24.0,<25.0
ruff>=0.1,<1.0

# Type checking
mypy>=1.8,<2.0
django-stubs>=4.2,<6.0
djangorestframework-stubs>=3.14,<4.0
